+++ api/routers/__init__.py
@@ -1,1 +1,1 @@
-from .routes import router
+from .routers import router
